requirements.txt
Pillow>=10.0.0
mypy>=1.0.0
types-Pillow>=10.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.0